Auth Refactoring
Massive changes to Auth made. Added ReactObserver to monitor changes in AuthState. This then triggers a refresh of state in the AuthContext which is async for the username (pain in the arse). The context then drives redirection in the App component, removing the need to access context to redirect in components which are children to the app